# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bond Settlement System API
  description: REST API for SWIFT bond settlement instruction management
  version: 1.0.0
  contact:
    name: Settlement Team

servers:
  - url: http://localhost:9080/settlement
    description: WebSphere local development

paths:
  /api/settlement:
    post:
      summary: Create a new settlement instruction
      description: Submits a settlement instruction, builds MT541, and sends to SWIFT via IBM MQ
      operationId: createSettlement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettlementRequest'
            example:
              isin: "US0378331005"
              settlementDate: "2026-05-10"
              quantity: 1000000.00
              counterparty: "Goldman Sachs"
              bicCode: "GOLDUS33XXX"
              direction: "BUY"
              accountId: "ACC-001"
      responses:
        '201':
          description: Settlement instruction created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettlementResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: List settlement instructions
      description: Returns a paginated list of settlement instructions
      operationId: listSettlements
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
          description: Page number (0-based)
        - name: size
          in: query
          schema:
            type: integer
            default: 20
          description: Page size
      responses:
        '200':
          description: Paginated list of settlements
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedSettlementResponse'

  /api/settlement/{tradeRef}:
    get:
      summary: Get settlement instruction by trade reference
      operationId: getSettlement
      parameters:
        - name: tradeRef
          in: path
          required: true
          schema:
            type: string
          description: Unique trade reference identifier
      responses:
        '200':
          description: Settlement instruction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettlementResponse'
        '404':
          description: Settlement instruction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/holdings:
    get:
      summary: Get bond holdings
      operationId: getHoldings
      parameters:
        - name: accountId
          in: query
          required: false
          schema:
            type: string
          description: Filter by account ID
      responses:
        '200':
          description: List of bond holdings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HoldingResponse'

components:
  schemas:
    SettlementRequest:
      type: object
      required:
        - isin
        - settlementDate
        - quantity
        - counterparty
        - bicCode
        - direction
        - accountId
      properties:
        isin:
          type: string
          pattern: '^[A-Z]{2}[A-Z0-9]{9}[0-9]$'
          description: International Securities Identification Number
          example: "US0378331005"
        settlementDate:
          type: string
          format: date
          description: Settlement date (must be today or future)
          example: "2026-05-10"
        quantity:
          type: number
          format: decimal
          minimum: 0.01
          description: Number of securities
          example: 1000000.00
        counterparty:
          type: string
          maxLength: 100
          description: Counterparty name
          example: "Goldman Sachs"
        bicCode:
          type: string
          pattern: '^[A-Z]{6}[A-Z0-9]{2}([A-Z0-9]{3})?$'
          description: BIC/SWIFT code of counterparty
          example: "GOLDUS33XXX"
        direction:
          type: string
          enum: [BUY, SELL]
          description: Trade direction
          example: "BUY"
        accountId:
          type: string
          maxLength: 50
          description: Internal safekeeping account identifier
          example: "ACC-001"

    SettlementResponse:
      type: object
      properties:
        tradeRef:
          type: string
          description: Unique trade reference
          example: "TR-A1B2C3D4E5F6"
        isin:
          type: string
          example: "US0378331005"
        settlementDate:
          type: string
          format: date
          example: "2026-05-10"
        quantity:
          type: number
          example: 1000000.00
        counterparty:
          type: string
          example: "Goldman Sachs"
        bicCode:
          type: string
          example: "GOLDUS33XXX"
        direction:
          type: string
          enum: [BUY, SELL]
          example: "BUY"
        status:
          type: string
          enum: [PENDING, SENT, MATCHED, FAILED, CANCELLED]
          example: "SENT"
        accountId:
          type: string
          example: "ACC-001"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    HoldingResponse:
      type: object
      properties:
        accountId:
          type: string
          example: "ACC-001"
        isin:
          type: string
          example: "US0378331005"
        quantity:
          type: number
          example: 5000000.00
        updatedAt:
          type: string
          format: date-time

    PagedSettlementResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/SettlementResponse'
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        totalElements:
          type: integer
          example: 42
        totalPages:
          type: integer
          example: 3

    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 400
        message:
          type: string
          example: "Validation failed"
        errors:
          type: array
          items:
            type: string
          example: ["isin: Invalid ISIN format"]
        timestamp:
          type: string
          format: date-time